+++ src/main/java/io/wcm/handler/mediasource/dam/impl/ngdm/WebOptimizedImageDeliveryServiceImpl.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package io.wcm.handler.mediasource.dam.impl.dynamicmedia;
+package io.wcm.handler.mediasource.dam.impl.ngdm;
 
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
